Fixed RTEs from not setting pointers. Lobby.AddChannel now broadcasts LobbyPlayerCount. New test for leaving quickplay.
